Add packages
- Add Duke package for main file
- Add Duke.util package for util files
