--- src/main/java/DukeActions.java
@@ -1,8 +1,0 @@
-import java.util.ArrayList;
-import java.util.Map;
-
-
-public interface DukeActions {
-
-    public boolean run (Map<String, String> map, DukeTaskList list, DukeDB database, DukeDateConfig config, Ui ui) throws DukeException;
-}
