+++ src/main/java/duke/DukeActions.java
@@ -1,0 +1,13 @@
+package duke;
+
+import duke.util.DukeConfig;
+import duke.util.DukeTaskList;
+import duke.util.Ui;
+
+import java.util.Map;
+
+
+public interface DukeActions {
+
+    public boolean run (Map<String, String> map, DukeTaskList list, DukeDB database, DukeConfig config, Ui ui) throws DukeException;
+}
