+++ src/main/java/duke/DukeCommands.java
@@ -1,3 +1,10 @@
+package duke;
+
+import duke.util.DukeConfig;
+import duke.util.DukeTaskList;
+import duke.util.Parser;
+import duke.util.Ui;
+
 import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.Map;
@@ -6,15 +13,15 @@ import java.util.Optional;
 public enum DukeCommands {
 
     BYE("bye",
-            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeDateConfig config, Ui ui) -> {
+            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeConfig config, Ui ui) -> {
                 Duke.printMsg("Bye. Hope to see you again soon!");
                 return false;
             }), LIST("list",
-            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeDateConfig config, Ui ui) -> {
+            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeConfig config, Ui ui) -> {
                 ui.printTaskList(list);
                 return true;
             }), DONE("done",
-            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeDateConfig config, Ui ui) -> {
+            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeConfig config, Ui ui) -> {
                 list.setDone(Parser.parseInt(map.get("done"))
                         .orElseThrow(() -> new DukeException("Positional argument 'index' is not valid.")) - 1)
                         .map(x -> {
@@ -25,11 +32,11 @@ public enum DukeCommands {
                         .orElseThrow(() -> new DukeException("The provided index is out of bounds."));
                 return true;
             }), EVENT("event",
-            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeDateConfig config, Ui ui) -> {
+            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeConfig config, Ui ui) -> {
                 if (!map.containsKey("/at")) {
                     throw new DukeException("Missing positional argument '/at'.");
                 } else if (map.get("event") == null || map.get("/at") == null) {
-                    throw new DukeException("Event body or /at body cannot be empty.");
+                    throw new DukeException("duke.Event body or /at body cannot be empty.");
                 } else {
                     Task event = new Event(map.get("event"),
                             map.get("/at"));
@@ -40,13 +47,13 @@ public enum DukeCommands {
                 }
                 return true;
             }), DELETE("delete",
-            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeDateConfig config, Ui ui) -> {
+            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeConfig config, Ui ui) -> {
                 list.toRemove(Parser.parseInt(map.get("delete"))
                         .orElseThrow(() -> new DukeException("The delete command is either missing the positional argument 'index' or it " + "is invalid.")) - 1)
                         .orElseThrow(() -> new DukeException("Failed to remove task from the tasklist."));
                 return true;
             }), DEADLINE("deadline",
-            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeDateConfig config, Ui ui) -> {
+            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeConfig config, Ui ui) -> {
                 if (!map.containsKey("/by")) {
                     throw new DukeException("Missing positional argument " + "'/by'.");
                 } else if (map.get("deadline") == null || map.get("/by") == null) {
@@ -64,9 +71,9 @@ public enum DukeCommands {
                 }
                 return true;
             }), TODO("todo",
-            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeDateConfig config, Ui ui) -> {
+            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeConfig config, Ui ui) -> {
                 if (map.get("todo") == null) {
-                    throw new DukeException("Todo body cannot be empty.");
+                    throw new DukeException("duke.Todo body cannot be empty.");
                 } else {
                     Task event = new Todo(map.get("todo"));
                     list.addTask(event)
@@ -76,7 +83,7 @@ public enum DukeCommands {
                 }
                 return true;
             }), INVALID("invalid",
-            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeDateConfig config, Ui ui) -> {
+            (Map<String, String> map, DukeTaskList list, DukeDB database, DukeConfig config, Ui ui) -> {
                 throw new DukeException("Unrecognized command.");
             });
 
@@ -85,7 +92,7 @@ public enum DukeCommands {
     final DukeActions action;
 
     /**
-     * Simple constructor to initialise a DukeCommands enum
+     * Simple constructor to initialise a duke.DukeCommands enum
      *
      * @param command The command that this instance serves.
      * @param action  The action that the command has.
