+++ src/main/java/duke/DukeDB.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.util.Parser;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
