+++ src/main/java/duke/util/DukeTaskList.java
@@ -1,3 +1,7 @@
+package duke.util;
+
+import duke.Task;
+
 import java.util.ArrayList;
 import java.util.Optional;
